move security logic into its own file
