+++ backend/api/token.py
@@ -1,0 +1,30 @@
+from datetime import timedelta
+from typing import Annotated
+
+from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordRequestForm
+from sqlalchemy.orm import Session
+
+from ..crud.security_logic import authenticate_user, ACCESS_TOKEN_EXPIRE_MINUTES, create_access_token
+from backend.db.config import get_db
+from ..schema.token import Token
+
+router = APIRouter()
+
+
+@router.post("/token")
+def login(
+        form_data: Annotated[OAuth2PasswordRequestForm, Depends()], db: Session = Depends(get_db)
+) -> Token:
+    user = authenticate_user(db, form_data.username, form_data.password)
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect username or password",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    access_token = create_access_token(
+        data={"sub": user.username}, expires_delta=access_token_expires
+    )
+    return Token(access_token=access_token, token_type="bearer")
