+++ backend/main.py
@@ -1,78 +1,16 @@
-from datetime import datetime, timedelta, timezone
-
 import uvicorn
-from fastapi import Depends, FastAPI, HTTPException, status
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
-from jose import JWTError, jwt
-from typing import Annotated
-
-from sqlalchemy.orm import Session
 
 from backend.api.api import api_router
-from backend.crud import crud_user
-from backend.crud.security_logic import *
 from backend.db import db_user, db_article
-from backend.db.config import engine, get_db
-from backend.schema.user import User
-from backend.schema.token import Token, TokenData
+from backend.db.config import engine
 
 # This creates DB schema according to the models localed in the "db" folder:
 db_user.Base.metadata.create_all(bind=engine)
 db_article.Base.metadata.create_all(bind=engine)
 
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
-
 app = FastAPI()
-
-
-def authenticate_user(db, username: str, password: str):
-    user = crud_user.get_user(db, username)
-    if not user:
-        return False
-    if not verify_password(password, user.password):
-        return False
-    return user
-
-
-def create_access_token(data: dict, expires_delta: timedelta | None = None):
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.now(timezone.utc) + expires_delta
-    else:
-        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
-
-
-def get_current_user(db, token: Annotated[str, Depends(oauth2_scheme)]):
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        token_data = TokenData(username=username)
-    except JWTError:
-        raise credentials_exception
-    user = crud_user.get_user(db, username=token_data.username)
-    if user is None:
-        raise credentials_exception
-    return user
-
-
-def get_current_active_user(
-        current_user: Annotated[User, Depends(get_current_user)]
-):
-    if current_user.disabled:
-        raise HTTPException(status_code=400, detail="Inactive user")
-    return current_user
-
 
 origins = [
     "http://localhost:3000"
@@ -93,24 +31,6 @@ def root():
     a = "a"
     b = "b" + a
     return {"hello world": b}
-
-
-@app.post("/token")
-def login(
-        form_data: Annotated[OAuth2PasswordRequestForm, Depends()], db: Session = Depends(get_db)
-) -> Token:
-    user = authenticate_user(db, form_data.username, form_data.password)
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Incorrect username or password",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-    access_token = create_access_token(
-        data={"sub": user.username}, expires_delta=access_token_expires
-    )
-    return Token(access_token=access_token, token_type="bearer")
 
 
 app.include_router(api_router)
